- Fix various compiler warnings - Ignore compiler warnings in Boost headers
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13342 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/lts/include/mcrl2/lts/probabilistic_label.h
@@ -107,7 +107,7 @@ class probabilistic_label
       assert(enumerator.size() <= denominator.size());
       
       // Check whether the denominator fits in number_t.
-      if (denominator.size() >= std::numeric_limits<number_t>::digits10)
+      if (denominator.size() >= (size_t)std::numeric_limits<number_t>::digits10)
       {
         std::string shortened_denominator = denominator.substr(0, std::numeric_limits<number_t>::digits10);
         m_denominator = to_number_t(shortened_denominator);
